image_resizer/requirements.txt
pika==1.3.1
Pillow==9.4.0
